Trait quick_js::Callback

pub trait Callback<F>: RefUnwindSafe {
    // Required methods
    fn argument_count(&self) -> usize;
    fn call(
        &self,
        args: Vec<JsValue>
    ) -> Result<Result<JsValue, String>, ValueError>;
}

The Callback trait is implemented for functions/closures that can be used as callbacks in the JS runtime.

Required Methods

fn argument_count(&self) -> usize

The number of JS arguments required.

fn call( &self, args: Vec<JsValue> ) -> Result<Result<JsValue, String>, ValueError>

Execute the callback.

Should return:

Err(_) if the JS values could not be converted

Ok(Err(_)) if an error ocurred while processing. The given error will be raised as a JS exception.

Ok(Ok(result)) when execution succeeded.